Import the ppoprf::ffi module from sta-rs.
We need C ffi bindings to call the randomness server implementation
in Rust from the Go service implementation.

For reproducibility we want to control the build environment and
dependency versions of the code, so we need a Cargo.lock. It's
simpler to maintain everything for that as part of this repository
rather than cloning and building from a submodule.

This version is based on the ppoprf-ffi branch from
https://github.com/NullHypothesis/sta-rs