Update go code to build against local rust ffi.
Instead of checking out a revision of the upstream repo containing
the ffi bindings to the rust ppoprf implementation which is the
core of the randomness server, keep the bindings in this repository
and build against the upstream ppoprf crate as a git dependency.

Since we want reproducible builds, having the Cargo.lock alongside
go.sum in a single repo simplifies dependency tracking.

Also expand the library name a bit so it's more comprehensible.