Make epoch loop start at next epoch.
Odds are that the server's start time does not coincide with the next
epoch's start time.  We therefore wait until the next epoch begins, at
which point we begin our epoch rotation loop.

To make this work, this commit also modified getEpoch(), and taught it
to take as input the epoch length.