Use the current epoch if none is specified.
If the client request doesn't specify an epoch, the json decoder
uses the default value of zero. Instead, store the epoch as a
pointer, so we can use the `nil` value to encode the absence of
the key in the request body.

When this epoch pointer in `nil` substitute the current epoch
so the request works as might be expected.