Remove lock and reduce protected code of another.
In issue https://github.com/brave-experiments/star-randsrv/issues/35
we found a lock that's not necessary.  This commit removes that lock.

The commit also reduces the code that's protected by another lock.
(However, we still need to protect the creation of the srvInfoResponse
struct because we're reading srv.pubKey which may simultaneously be
written by the init function.)