Fix up logging
Use the `info!` convenience macro with the default logger,
which reads the `RUST_LOG` environment variable. So run with:

    RUST_LOG=tower_http=trace,star_randsrv=debug cargo run

for reasonable output.

By default it's `ERROR` level only, so there's no convenience
info about requests or listening port; would be nice to default
to `star_randsrv=info` or something like that.