Rotate OPRF key when epoch list is exhausted.
Instead of leaving all epochs puctured and later panicking, create
and fresh OPRFServer instance with the same set of epochs and
continue with that.

Abstract the struct initialization into OPRFServer::new() so it
can be called from both locations, Move the `future_epochs`
initialization inside the background loop so the same code
can be used each time a new state is initialized.