Fix OPRF key rotation
The previous code reset the `future_epochs` vector before the
empty check that was supposed to trigger re-initialization.
So OprfServer.epoch was updated, but the actual set of viable
epochs was punctured, resulting in perpetual "No prefix found"
errors.

Instead, assume we'll use contiguous blocks of epoch numbers
and store them as a Range, which is smaller and more convenient.
`EpochRange` never needs to change and can be passed to
`OprfServer::new` each time the key is rotated.

Note that because `Range` and `InclusiveRange` are different
types, we can't both keep the range as `u8` and include epoch
255. I though code simplicity was more important here.