Consolidate request building.
Move the redundant parts of generating a request for an endpoint
into a helper function. If a payload is included, the request
is POSTed as application/json.

This doesn't save any code really, but makes the tests a bit
easier to read since we lose the conversion code to build the
request.