Test specifying an epoch
The /randomness endpoint allows specifying a specific epoch tag
to support the same route offering multiple valid epochs at the
same time. For example this could be two series with different
cadences or an overlap to prevent failed queries close to the
rotation time.

Verify that the current epoch can be accepted, but that earlier
and later epochs are rejected.