Allow clippy::assertions_on_constants in tests::epoch
The epoch test has assertion that the epoch test values it uses
are distinct from the accepted value. These compile to nothing
because they are const expressions. However, it's nice to have
the assertion in case the base constant changes in an incompatible
way.

`cargo clippy --all-targets` warns about this test, so silence
the lint for a clean report while keeping the assertion.