Use the Amazon Web Services container registry
This fetches the Docker image from Amazon's container registry instead of Docker's, which prevents the frequent rate limits in our reproducible build pipeline. At the time of commit, at least, the `docker/library` path is a mirror of the docker.io registry with identical images hashes for these
three tags.

Co-authored-by: Philipp Winter <phw@brave.com>