Update Makefile
Adjust Makefile to build a versioned eif file in docker, which
shortcuts for deployment. Some of this is style changes, but
other parts are functional; the previous `ko` build didn't
work for the non-go parts of the container.

This reduces the memory cap to 512 MB to match the default
allocation limit. This simplifies testing, especially since
bumping the limit has been failing occasionaly in my testing.
It's certainly enough memory for verifying deployment, but
benchmarking is needed to see if it's sufficient to handle
a full network load.