Check for overflow incrementing the epoch.
The ppoprf epoch tag is a `u8`. When using the full epoch range
0..=255, the final increment would overflow. In debug builds this
would panic. In release builds, it would roll over unchecked, then
panic at the end of the next interval when it tried to puncture
the already-punctured starting epoch value. In either case, the
panic happened while the write lock was held, poisoning it.

The tokio thread runner would catch the panic, but with the
`RwLock` poisoned, it couldn't respond to further queries.

Instead, check for overflow and use that to trigger key rotation,
just as we do when the epoch is out of the configure range. This
is more localized than promoting the epoch counter and range to
u16 and narrowing at the ppoprf calls.