Speed up container builds
Change the Dockerfile to only copy relevant source files into
the each build context. The greatly speeds up development builds
where local artefacts (especially the rust `target` subtree) can
add GB to the package passed to the builder. This is more verbose
and possibly brittle, but it allows faster iteration.

Also use the sparse+https protocol to query crates.io. This is
supported in Rust 1.68.0 and later, and is expected to be the
default starting with 1.70.0. This just downloads metadata for
referenced crates instead of the complete index, which is much
faster for one-off builds like a container image.