Align the epoch schedule with --epoch-basetime
When the `--epoch-basetime` argument is present and parses to
a time in the past, calculate which epoch  would be current
if the server had been running since that time. This helps
align epoch numbers of rotation points between invocations
which can be less disruptive.

Calculate the delay until the next rotation point each time
through the loop. Converting through seconds was the best
path I could find to align `tokio::time` with calendar time.