github actions: Remove codeql config
We enabled this per the security checklist, but CodeQL still doesn't
support the Rust language, and recently has started reporting failures
because it's confused by finding yaml but not golang code to audit.

Just remove the config to reduce noise; we can add it back if we
add code it can audit.