Update tests for axum 0.7
This release removes the re-export of hyper::body, which has moved
to the http-body crate. Worse, the `to_bytes` method we were using
wasn't stablized with the 1.0 release of that crate, and was instead
moved to a new http-body-util crate.

axum 0.7.2 will likely add a native Body::to_bytes method we can
use for this. In the meantime use a similar conversion ourselves.
See https://github.com/tokio-rs/axum/pull/2373