Remove direct dependency on metrics_exporter_prometheus
We referenced this just to import the `PrometheusHandle` type
returned by `axum_prometheus::PrometheusMetricLayer::pair()`.
That crate re-exports its version of that same crate, so we
can refer to it by that path and not need to keep them in sync.

Thanks to Péter Leéh for the suggestion in
https://github.com/Ptrskay3/axum-prometheus/issues/38