Use the non-slim rust base image for the deployment build
The non-slim container image includes both the `libssl-dev` and
`pkg-config` packages, which are required by the `openssl-sys`
crate dependency added by `axum-prometheus` 0.6, but not included
in the `build-essential` package we were installing for other
dependencies.

The image is almost twice the size but includes all of build-esential
as well, so it may well run faster.

Probably we can go back to slim if the `hypertls` dep can be
dropped from the prometheus stack's push feature, which we're
not using.